Mise a jour des ressources

+++ book/src/main/java/com/BibliothequeP7/book/dao/BookDao.java
@@ -3,5 +3,5 @@ package com.BibliothequeP7.book.dao;
 import com.BibliothequeP7.book.entities.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookDao extends JpaRepository<Book, Integer> {
+public interface BookDao extends JpaRepository<Book, Long> {
 }
